refactor: rename core to production 1
