--- cognite/powerops/resync/config_classes/marked_configuration/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import annotations
-
-from cognite.powerops.resync.config_classes.marked_configuration._core import PriceScenario, PriceScenarioID
-from cognite.powerops.resync.config_classes.marked_configuration.benchmarking import BenchmarkingConfig
