+++ cognite/powerops/resync/config_classes/market/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+from cognite.powerops.resync.config_classes.market._core import PriceScenario, PriceScenarioID
+from cognite.powerops.resync.config_classes.market.benchmarking import BenchmarkingConfig
