+++ cognite/powerops/resync/config_classes/market/_core.py
