+++ cognite/powerops/resync/config_classes/market/dayahead.py
@@ -7,8 +7,8 @@ from typing import ClassVar, Dict, List, Optional
 
 from pydantic import BaseModel, Field, field_validator
 
-from cognite.powerops.resync.config_classes.marked_configuration import PriceScenarioID
-from cognite.powerops.resync.config_classes.marked_configuration._core import Configuration, RelativeTime
+from cognite.powerops.resync.config_classes.market import PriceScenarioID
+from cognite.powerops.resync.config_classes.market._core import Configuration, RelativeTime
 
 
 class BidMatrixGeneratorConfig(BaseModel):
