+++ cognite/powerops/resync/config_classes/production/connections.py
