+++ cognite/powerops/resync/config_classes/production/generator.py
