+++ cognite/powerops/resync/config_classes/production/plant.py
