+++ cognite/powerops/resync/config_classes/production/watercourse.py
